+++ src/components/SearchArea/index.tsx
