+++ src/components/SearchAreaWrap/index.tsx
